Remove Search and do refactoring
