+++ package.json
@@ -20,7 +20,7 @@
     "@angular/router": "3.1.0",
     "angular2-apollo": "0.5.0",
     "angular2-emojify": "0.0.1",
-    "angular2-moment": "1.0.0-beta.5",
+    "angular2-moment": "1.0.0-beta.6",
     "apollo-client": "0.4.22",
     "core-js": "2.4.1",
     "rxjs": "5.0.0-rc.1",
