+++ src/app/feed/feed-entry.component.ts
@@ -1,5 +1,5 @@
 import {Component, Input, Output, EventEmitter} from '@angular/core';
-import {Entry, User} from '../../../schema-types';
+import {Entry, User} from '../../schema-types';
 
 export interface OnVoteEvent {
   repoFullName: string;
