+++ src/app/new-entry/new-entry.component.html
