+++ src/app/new-entry/new-entry.model.ts
