+++ src/app/profile/profile.component.ts
@@ -1,6 +1,7 @@
 import {Component, OnInit, OnDestroy} from '@angular/core';
 import {Angular2Apollo} from 'angular2-apollo';
 import {Subscription} from 'rxjs/Subscription';
+
 import {CurrentUserQuery} from './profile.model';
 
 @Component({
