+++ src/app/shared/index.ts
@@ -1,9 +1,8 @@
 import { InfoLabelComponent } from './info-label.component';
 import { LoadingComponent } from './loading.component';
 import { RepoInfoComponent } from './repo-info.component';
-import { Component } from '@angular/core';
 
-export const SHARED_DECLARATIONS: Component[] = [
+export const SHARED_DECLARATIONS: any[] = [
   InfoLabelComponent,
   LoadingComponent,
   RepoInfoComponent
