+++ src/app/shared/loading.component.ts
